Merged revisions 43710 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
................
r43710 | russell | 2006-09-26 16:56:42 -0400 (Tue, 26 Sep 2006) | 17 lines

(This was actually BE-65)

Merged revisions 43708 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.2

........
r43708 | russell | 2006-09-26 16:49:21 -0400 (Tue, 26 Sep 2006) | 7 lines

Back in revision 4798, this message was changed from using ast_cli() to directly
calling write().  During this change, checking if this was a remote console was
removed.  This caused this message about using "exit" or "quit" to exit an
Asterisk console to come up in times where it did not make sense.  This change
restores the check to see if this is a remote console before printing the
message.  (fixes BE-4)

........

................


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@43715 f38db490-d61c-443f-a65b-d21fe96a405b