Blocked revisions 43756 via svnmerge
........
r43756 | russell | 2006-09-27 00:35:18 -0400 (Wed, 27 Sep 2006) | 10 lines

Backport revision 43754 from the trunk, which removes an unused buffer from
mm_login to close bug 8038, as well as addresses some formatting and coding
guidelines issues in passing.

Originally, I did not commit this to 1.4 since it is not necessarily fixing a
bug.  However, since the IMAP storage code is brand new, I decided it would
be better to make the change here as well, in case someone has to work on this
code to address issues in the very near future.  I don't want to make
unnecessary merge problems going to the trunk.

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@43757 f38db490-d61c-443f-a65b-d21fe96a405b