Merged revisions 46776 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r46776 | russell | 2006-11-01 13:24:17 -0500 (Wed, 01 Nov 2006) | 9 lines

soxmix and Asterisk expect different file extensions for certain formats.  This
was already handled for the wav49 format.  However, it was not handled for
ulaw and alaw.  I fixed this in such a way that using the alternate extensions
for ulaw and alaw will only happen if we know we're calling soxmix, and not a
custom script defined using the MONITOR_EXEC variable.  The wav49 processing
was left alone so that external scripts will see no behavior change.
(issue #7550, reported by mnicholson, proposed patch by junky, committed fix
is a bit different)

........


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@46778 f38db490-d61c-443f-a65b-d21fe96a405b