Explain the use device status system implemented in SIP for subscriptions, queues and manager a bit better.
Like in 1.2, you will get more detailed information if you set a call 
limit for a device. When the call limit is reached, the status system will
report a device as busy.

For queues, setting a call limit per SIP device is propably a requirement.

In most cases, it will work much better if you only use type=peer and not
type=friend. We might decide to backport the new setting from trunk to
apply all call limits to the peer part of a friend only.



git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@48113 f38db490-d61c-443f-a65b-d21fe96a405b