Blocked revisions 48246 via svnmerge
........
r48246 | qwell | 2006-12-04 15:20:34 -0600 (Mon, 04 Dec 2006) | 7 lines

Revert change from 8016 - this breaks other stuff...  Needs further review.

Tip: When you've reported a bug about something and somebody has put up a
patch for it..  It's not a good idea to open a completely new bug and say that
something is broken because of the patch in the other bug - PLEASE mention
something in the bug where the patch was actually created.

........


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@48247 f38db490-d61c-443f-a65b-d21fe96a405b