Merged revisions 48252 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
................
r48252 | tilghman | 2006-12-04 19:34:34 -0600 (Mon, 04 Dec 2006) | 14 lines

Merged revisions 48251 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.2

........
r48251 | tilghman | 2006-12-04 19:26:08 -0600 (Mon, 04 Dec 2006) | 6 lines

If the recording in the database is too large, it will fail to retrieve with
an mmap error.  Not too sure why this doesn't happen when we put it in the
database, also, but since that doesn't seem to be broken, I'm not going to fix
it (at least until someone reports it).  Solution is to ask for the file in
smaller chunks.  (Bug 8385)

........

................


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@48253 f38db490-d61c-443f-a65b-d21fe96a405b