This bug was fixed in the trunk by the recent work to consolidate the various string handling into a single API, so this revision is blocked here.
Blocked revisions 49388 via svnmerge

........
r49388 | russell | 2007-01-03 23:33:00 -0500 (Wed, 03 Jan 2007) | 6 lines

Fix the REALTIME() dialplan function.  ast_build_string() advances the string
pointer to the position to begin the next write into the buffer.  So, this 
pointer can not be used to copy the contents of the string later.  The
beginning of the buffer must be saved.  Interestingly enough, this code could
not have ever worked.  (Pointed out by Sebb on IRC, thanks!)

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@49389 f38db490-d61c-443f-a65b-d21fe96a405b