Merged revisions 58894 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r58894 | russell | 2007-03-14 11:33:01 -0500 (Wed, 14 Mar 2007) | 8 lines

By default, don't attempt to do any CallerID handling at all with SLA because
it is known to not work properly in some situations.  However, add an option to
enable it for those that would like to use it anyway.

The short story behind this is that to properly handle CallerID with SLA, we
need the ability to change the CallerID on an existing call, and we are not
ready to handle that.

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@58895 f38db490-d61c-443f-a65b-d21fe96a405b