Merged revisions 59341 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r59341 | russell | 2007-03-29 11:55:39 -0500 (Thu, 29 Mar 2007) | 8 lines

When the IAX2 read callback gets called, return NULL instead of a "null frame".
This will cause Asterisk to hangup the call instead of keep trying whatever it
was doing.  Under normal conditions, this function would *never* be called.
However, the author of this patch says an error will occur that will cause it
to get called every 100 thousand calls or so.  When this does happen, it puts
the channel in a loop that eventually brings down the system.  So, hangup up
the call is certainly a better alternative.  (issue #8286, john)

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@59343 f38db490-d61c-443f-a65b-d21fe96a405b