Merged revisions 67993 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r67993 | oej | 2007-06-07 11:00:44 +0200 (Thu, 07 Jun 2007) | 6 lines

Issue #9738 - Make sure we can unload res_jabber. Patch by phsultan - thanks!

Due to a bug in the iksemel library, this will not work if you are using GTLS
in the connection. That's being investigated. If you figure out a way to handle
that without us having to patch iksemel, let us know in the bug report. Thanks.

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@68026 f38db490-d61c-443f-a65b-d21fe96a405b