Merged revisions 74766 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r74766 | russell | 2007-07-11 17:53:26 -0500 (Wed, 11 Jul 2007) | 5 lines

The function make_trunk() can fail and return -1 instead of a valid new call
number.  Fix the uses of this function to handle this instead of treating it
as the new call number.  This would cause a deadlock and memory corruption.
(possible cause of issue #9614 and others, patch by me)

........


git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@74767 f38db490-d61c-443f-a65b-d21fe96a405b