Merged revisions 77410 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r77410 | russell | 2007-07-26 16:23:23 -0500 (Thu, 26 Jul 2007) | 10 lines

AST_DEVMODE was defined in trunk, but not in 1.4.  When Asterisk is compiled
under dev mode, AST_DEVMODE will get defined in buildopts.h.  Change 1.4 to
define it in the same way that trunk does.  Also, revert the change that added
this define in the Makefile

The advantage to doing it this way is that buildopts.h gets installed when
you install Asterisk.  Then, when building any out of tree modules, or
building asterisk-addons, these modules know which options the rest of Asterisk
was built with.

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@77411 f38db490-d61c-443f-a65b-d21fe96a405b