Merged revisions 77778 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r77778 | file | 2007-07-30 14:11:02 -0300 (Mon, 30 Jul 2007) | 4 lines

(closes issue #10327)
Reported by: kkiely
Instead of directly mucking with the extension/context/priority of the channel we are transferring when it has a PBX simply call ast_async_goto on it. This will ensure that the channel gets handled properly and sent to the right place.

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@77779 f38db490-d61c-443f-a65b-d21fe96a405b