Blocked revisions 80497 via svnmerge
........
r80497 | russell | 2007-08-23 11:53:52 -0500 (Thu, 23 Aug 2007) | 5 lines

This is a hack to maintain old behavior of chan_iax2.  This ensures that if
the peers and users are being stored in a linked list, that they go in the
list in the same order that the older code used.  This is necessary to maintain
the behavior of which peers and users get matched when traversing the container.

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@80498 f38db490-d61c-443f-a65b-d21fe96a405b