Merged revisions 81952 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r81952 | russell | 2007-09-08 08:42:26 -0500 (Sat, 08 Sep 2007) | 11 lines

(closes issue #10672)
Bump the cleancount so that a "make clean" will be forced.  This is needed
because my fix in revision 81599 made a change to a data structure in file.h,
and since file dependency tracking is only on with dev-mode enabled, file
format modules that don't get rebuilt may crash, as is the case with this issue.

This makes me wonder - how much faster does the code build without the file
dependency tracking enabled?  If it doesn't make much of a difference, then it
may be worth just keeping it on all of the time, or perhaps just not in release
tarballs, so that this type of issue is avoided.

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@81953 f38db490-d61c-443f-a65b-d21fe96a405b