Merged revisions 85604 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r85604 | russell | 2007-10-15 11:54:57 -0500 (Mon, 15 Oct 2007) | 6 lines

Make the default for the srvlookup option to be yes.  It doesn't really make
sense for it to default to off.  The default configuration file has it on, and
proper RFC behavior, as indicated by a comment in the code, is for it to be on.
So, let's have it on by default to make lives easier.
(closes issue #10954, suggested by jtodd)

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@85605 f38db490-d61c-443f-a65b-d21fe96a405b