Merged revisions 92158 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r92158 | oej | 2007-12-10 15:04:44 +0100 (MÃ¥n, 10 Dec 2007) | 16 lines

Avoid reinvite race situations with two Asterisks trying
to reinvite each other in 1.4 and trunk. 

This patch implements support for the 491 error code that
Asterisk 1.4 generates on situations where we get an 
incoming INVITE and already has one in progress.

Thanks to mavetju for reporting and to Raj Jain for an
excellent explanation of the problem.

Patch by myself. Tested with 8 Asterisk servers connected
to each other in a training network.

Closes issue #10481


........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@92159 f38db490-d61c-443f-a65b-d21fe96a405b