Blocked revisions 93000 via svnmerge
........
r93000 | russell | 2007-12-14 11:36:08 -0600 (Fri, 14 Dec 2007) | 7 lines

There are a lot of existing systems that #include non-existent files.  So, to
make the transition to treating this as an error a bit less painless, just issue
a huge error message for now.  Then, later, we can reinstate the code that treats
it as a failure.

(Thanks to philippel for the feedback)

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@93001 f38db490-d61c-443f-a65b-d21fe96a405b