Merged revisions 96575 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r96575 | tilghman | 2008-01-04 17:03:40 -0600 (Fri, 04 Jan 2008) | 7 lines

Fix the problem of notification of a device state change to a device with a '-'
in the name.  Could probably do with a better fix in trunk, but this bug has
been open way too long without a better solution.
Reported by: stevedavies
Patch by: tilghman
(Closes issue #9668)

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@96576 f38db490-d61c-443f-a65b-d21fe96a405b