Merged revisions 106704 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r106704 | russell | 2008-03-07 11:16:58 -0600 (Fri, 07 Mar 2008) | 8 lines

Change a warning message to a debug message.  This is happening quite frequently,
and it is not worth spamming users with these messages unless we are pretty confident
that it should never happen.  As it stands today, it _will_ and _does_ happen and
until that gets cleaned up a reasonable amount on the development side, let's not
spam the logs of everyone else.

(closes issue #12154)

........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@106707 f38db490-d61c-443f-a65b-d21fe96a405b