Merged revisions 110635 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r110635 | mmichelson | 2008-03-25 10:40:33 -0500 (Tue, 25 Mar 2008) | 7 lines

When reverting a commit, I accidentally left in this bit which was an experiment
to see what would happen. It passed the compile test, and I didn't notice I had
left this change in too.

So this is a revert of a revert...sort of.


........


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@110636 f38db490-d61c-443f-a65b-d21fe96a405b