(closes issue #12362) [redo of 113012]
This fixes a for loop (in realtime_peer) to check all the ast_variables the loop was intending to test rather than just the first one. The change exposed the problem of calling memcpy on a NULL pointer, in this case the passed in sockaddr_in struct which is now checked. 



git-svn-id: http://svn.digium.com/svn/asterisk/branches/1.4@113240 f38db490-d61c-443f-a65b-d21fe96a405b